Fix Ktlint format after global find and replace
When global find and replace is invoked, this also triggers the ActionOnSave. It seems that the global find and replace does change the "document.text" but not the "psiFile.text" before the save is completed. If KtlintFormat then uses the "psiFile.text" (e.g. the original code) it undoes the change made by the global find and replace.