Do not show violations which can be autocorrected / add intention to format file
Violations which can be autocorrected should never be displayed. They are distracting, and it could lead to developers starting to fix the issues manually. A (weak) warning is displayed, with the number of errors that can be autocorrected. On this warning, an intention is shown to format the file. This intention is also available in case the plugin is not enabled.