Add action and intention to show all ktlint violations in the editor window
The EditorNotificationProvider which displays a notification panel at top of the editor is no longer needed. The annotator now displays a file level annotation when the plugin is not set enabled. This banner is only displayed whenever ktlint actually finds a violation.

Add action and intention to show all ktlint violations in the editor window

The EditorNotificationProvider which displays a notification panel at top of the editor is no longer needed. The annotator now displays a file level annotation when the plugin is not set enabled. This banner is only displayed whenever ktlint actually finds a violation.