Minimize the "ktlint" ShadowJar
When creating a shadow jar of a specific ktlint ruleset all transitive dependencies are included in that jar. Combining the individual ruleset shadow jars in the plugin-jar blows up the size of the plugin jar as the same dependencies are included multiple times. By first combining the ruleset shadow jars into another shadowed jar the duplicate dependencies are pruned.

Closes #498