Support publication of dev versions (#517)
A dev-version is published on the new dev channel to the plugin. As plugin developer it becomes easier to keep using the last plugin version as it can be installed from Jebrains marketplace.

The beta and dev version number are automatically expanded with a build timestamp to create unique and identifiable version numbers.