Fix loading custom ruleset
This resolves the exception about missing the RulesetProviderV3 class when loading a custom ruleset.

Also, ignore the Standard rules which are provided via the custom ruleset. The plugin allows the user to select a specific version of the standard rules provided via ktlint. Loading the rules both from Ktlint and via the custom ruleset provider might lead to conflicts. As of that, the standard rules provided via the custom ruleset provider are ignored.