Change platform type from IC to IU (#716)
Change platform type as starting from Idea 2025.3 the IC (Community Edition) build no longer exists and as of that cannot be verified. PlatformType and versions are now set explicitly in the pluginVerifier step to workaround this. The Intellij Plugin Template has not yet catched up with this problem. So in the future, it has to be aligned with the template.