Assign notification to different notification groups for better customizability by user (#754)
Configuration and rule notification are now displayed in sticky balloons by default. When switching projects, the notification about an invalid ktlint version configured in the "ktlint-version.properties" file is displayed only once.

Remaining notification are displayed in normal balloons which disappear automatically.

KtlintParseExceptions are suppressed. Those exceptions happen a lot when code is saved while it can not yet be compiled successfully.